move a few more
